fix cutadapt minlen to 32

+++ src/pipecraft-core/service_scripts/cut_primers_paired_end_reads.sh
@@ -19,15 +19,15 @@
 ##########################################################
 
 #load variables
-extension=${fileFormat}  # KEEP THIS (removed in some other scripts)
-mismatches=$"-e ${mismatches}"
-min_length=$"--minimum-length ${min_seq_length}"
+extension=${fileFormat}             # KEEP THIS (removed in some other scripts)
+mismatches=$"-e ${mismatches}"      # number of mismatches in primer search
+min_length=$"--minimum-length 32"   # minimum len of the output sequence. FIXED to 32 (in order to avoid 0 len seqs) because cutadapt --minimum-length does not behave as expected
 overlap=$"--overlap ${min_overlap}"
 cores=$"--cores ${cores}"
-no_indels=$no_indels
-discard_untrimmed=$"TRUE" #currently only fixed to TRUE 
-seqs_to_keep=$seqs_to_keep
-pair_filter=$pair_filter
+# $no_indels                        # "--no-indels" disallows insertions and deletions in a primer
+discard_untrimmed=$"TRUE"           # fixed to TRUE 
+# $seqs_to_keep                     # all/keep_only_linked
+# $pair_filter                      # both/any
 fwd_tempprimer=$forward_primers
 rev_tempprimer=$reverse_primers
 
@@ -42,6 +42,13 @@ output_dir=$"/input/primersCut_out"
 if [[ $no_indels == "true" ]]; then
     indels=$"--no-indels"
 fi
+# if keep_only_linked, then linked primers are REQUIRED (default = optional)
+if [[ $seqs_to_keep == "keep_only_linked" ]]; then
+    required_optional=$"required"
+else
+    required_optional=$"optional"
+fi
+
 start=$(date +%s)
 ### Check if files with specified extension exist in the dir
 first_file_check
@@ -71,7 +78,8 @@ checkerror=$(seqkit seq --quiet -t dna -r -p tempdir2/fwd_primer.fasta >> tempdi
 check_app_error
 checkerror=$(seqkit seq --quiet -t dna -r -p tempdir2/rev_primer.fasta >> tempdir2/rev_primer_RC.fasta 2>&1)
 check_app_error
-# Make linked primers files
+
+# Make linked primer files
 i=1
 while read LINE; do
     fwd_primer=$(echo $LINE | grep -v ">")
@@ -84,7 +92,7 @@ while read LINE; do
                 :
             else
                 echo ">primer$i" >> tempdir2/liked_fwd_revRC.fasta
-                echo "$fwd_primer...$rev_primer_RC" >> tempdir2/liked_fwd_revRC.fasta
+                echo "$fwd_primer;$required_optional...$rev_primer_RC;$required_optional" >> tempdir2/liked_fwd_revRC.fasta
                 ((i=i+1))
             fi
         done < tempdir2/rev_primer_RC.fasta
@@ -102,7 +110,7 @@ while read LINE; do
                 :
             else
                 echo ">primer$i" >> tempdir2/liked_rev_fwdRC.fasta
-                echo "$rev_primer...$fwd_primer_RC" >> tempdir2/liked_rev_fwdRC.fasta
+                echo "$rev_primer;$required_optional...$fwd_primer_RC;$required_optional" >> tempdir2/liked_rev_fwdRC.fasta
                 ((i=i+1))
             fi
         done < tempdir2/fwd_primer_RC.fasta
@@ -185,7 +193,9 @@ while read LINE; do
         -p $output_dir/$inputR2.$extension \
         $output_dir/$inputR1.round1.$extension $output_dir/$inputR2.round1.$extension 2>&1)
         check_app_error
-        rm $output_dir/$inputR1.round1.$extension $output_dir/$inputR2.round1.$extension
+        if [[ $debugger != "true" ]]; then
+            rm $output_dir/$inputR1.round1.$extension $output_dir/$inputR2.round1.$extension
+        fi
 
     elif [[ $seqs_to_keep == "keep_only_linked" ]]; then
         checkerror=$(cutadapt --quiet \
@@ -225,8 +235,9 @@ while read LINE; do
         -p $output_dir/$inputR2.$extension \
         $output_dir/$inputR1.round1.$extension $output_dir/$inputR2.round1.$extension 2>&1)
         check_app_error
-        rm $output_dir/$inputR1.round1.$extension
-        rm $output_dir/$inputR2.round1.$extension
+        if [[ $debugger != "true" ]]; then
+            rm $output_dir/$inputR1.round1.$extension $output_dir/$inputR2.round1.$extension
+        fi
     fi
 done < tempdir2/paired_end_files.txt
